Create README.md
The Bike Race Game in C++ is a console application game development project. It is interesting, simple to play and demonstrates the use of various features of C++ programming language and SDL. In order to enhance the aesthetic of the game, I have utilized the application of graphics in the game to make the game more realistic. The project is mainly focused on the use of SDL in C++; SDL is a popular 2D gaming graphics.

The source code of the Bike Race Game in C++ is written in C++ project format and is compiled in Code::Blocks IDE using GCC compiler. The source code of the game consists of eight user defined header files and a number of user defined functions  are within them to perform certain specific tasks.

Listed below are some of the important user-defined header files in Bike Race game:

variable.h: it manages the variables defining the different parameters of the game such as time, jumps, wall etc.
init.h:- This header file initializes the game and loads the images used in this bike game.
process.h:- This header file had been used in the game in order to control the playing mechanism of game.
coardinate.h:- For the movement of bike in the window, the coordinates needs to be defined which has been done using this header file in Bike Race Game in C++.
constants.h:- This header file is made to control the constants su.
ch as the height, width etc. of wall and other constants used in the source code.
collision.h: In the Bike Race Game in C++, this header file is for describing the collision of bike with the wall.
function.h:- All the user defined functions are included in this header file.
menu.h:-  The main menu of the game is displayed using this header file.
The Bike Race Game in C++ is very simple to start, play and terminate. In order to start the game, you should just click at the application file of the game and then click at New Game in the window.

Features:

Due the use of SDL, the game environment is colorful as well as beautiful.
You have three options in the main menu of the game, you can play new game or you can see the instructions or you can terminate the game without playing by clicking at exit.
The Bike Race Game in C++ is time restricted, you have to travel all the path within 1 min and 30 seconds without colliding with any wall. Once the bike is collides against the wall, you will lose the game.
The speed of bike is controlled by right navigation key.
You can use the Bike Race Game in C++ to learn use and control of graphics in C++. It can also be used to learn the basic syntaxes and different formats in C++ as well as file handling in C++. You can use this game as your semester project in C++.